Major flag-handling rewrite.
Code should be cleaner as most of things are done
in one place and aren't scattered. For example
policy checking is not done with argument parsing.

More set/get things are done through ppp interface
which allows to centralise policy checks and in future
to remove PPP_INTERNAL flag.

BTW. failure counting should work. Not tested.